university-hrm-backend-main/university-hrm-backend-main/pyproject.toml
[project]
name = "university-hrm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-settings>=0.2.3",
    "rich>=14.3.2",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
